feat: move toggle persistence from settings to tools page
